원 수정, render update 수정
각 mesh 의 update 부분에서 각자의 mesh 가 가지고 있는 Transform 을 업데이트 하지 않고 CRenderer.h 의 Transform 을 업데이트 하고 있었음 이것을 각 mesh 의 Update 에는 그에 해당하는 mesh의 Transform 을 연결하여 업데이트 되도록 한다.

+++ DirectX2DProejct/Engine_SOURCE/CCell.h
@@ -7,7 +7,7 @@ namespace sh
 		public CGameObject
 	{
 	private:
-		double degree;
+		float degree;
 		double radius;
 		sh::math::Vector3 position;
